Update Spline Meshes and level design.
Brick Landing circle is placed now, as well as a paved circle for the Sky Bridge, two types of Bollards, and a new spline to place street lamps. I think we will need to make multispline meshes out of all of this, after we figure that out. The bridge BP is a small example, but hopefully we can make out streets with paths, bollards, street lamps, and more out of one of these multispline mesh BPs. Check the example street next to the sky bridge for an Idea what I mean. We will have to change the scale of the road based on what vehicles we use too. It should be wide enough for two to pass each other on it.